+++ Accounting.Domain/ViewModels/AccrualViewModel.cs
@@ -4,6 +4,6 @@ namespace Accounting.Domain.ViewModels
     public class AccrualViewModel
     {
         public Guid EmployeeId { get; set; }
-        public int Ammount { get; set; }
+        public decimal Ammount { get; set; }
     }
 }
